requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.3
Django==3.2.6
django-allauth==0.48.0
django-cors-headers==3.8.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
oauthlib==3.2.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.3.0
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
wincertstore==0.2
